add method + small refactoring #pr
