+++ src/element/index.ts
@@ -63,3 +63,4 @@ export * from './waits/waitForElementToBeUnChecked';
 export * from './waits/waitForElementToBeChecked';
 export * from './waits/waitForElementToHaveAttribute';
 export * from './waits/waitForScopedSelector';
+export * from './actions/getElementsCollectionLength';
