+++ src/element/props/getClasses.ts
@@ -1,7 +1,11 @@
-import { getElement, SearchElementOptions } from '../getElement';
+import { getElement } from '../getElement';
 import { getProperty } from '../getProperty';
 import { DocumentContext } from '../../page';
-import { ElementPropertyType, SelectorOrElement } from '../types';
+import {
+  ElementPropertyType,
+  SearchElementOptions,
+  SelectorOrElement,
+} from '../types';
 
 /**
  * Returns an array of classes for the specified element.
