+++ src/element/waits/waitForElement.ts
@@ -1,8 +1,8 @@
 import { ElementHandle } from 'puppeteer-core';
 import { ACTION_TIMEOUT } from '../../settings';
-import { getElement, SearchElementOptions } from '../getElement';
+import { getElement } from '../getElement';
 import { DocumentContext } from '../../page';
-import { SelectorOrElement } from '../types';
+import { SearchElementOptions, SelectorOrElement } from '../types';
 import { WaitOptions } from '../../types';
 
 export async function waitForElement(
